Closes 25 - Changed  to

+++ include/indicators/details/stream_helper.hpp
@@ -102,8 +102,8 @@ public:
                       const std::string &lead, const std::string &remainder)
       : os(os), bar_width(bar_width), fill(fill), lead(lead), remainder(remainder) {}
 
-  std::ostream &write(float progress) {
-    auto pos = static_cast<size_t>(progress * static_cast<float>(bar_width) / 100.0);
+  std::ostream &write(size_t progress) {
+    auto pos = static_cast<size_t>(progress * bar_width / 100.0);
     for (size_t i = 0; i < bar_width; ++i) {
       if (i < pos)
         os << fill;
